Fix stop after each quarter run
I am resetting two variables for the issue that we had to re-start the
BMSim for each quarter. And let me explain it here.
-          Inside DesiredMk( ) function, it uses Current array to
calculate the amount of loans desired by credit quality and calculate
dollar amount. It is not cleaned up each quarter.
memset(Current + Bank, 0, sizeof(*Current)); inside LoansStartUp()
-          Inside MkeShare(), Rel_Pol is not cleaned up each quarter. It
carries the values from last quarter calculation.
memset(Rel_Pol, 0, sizeof(Rel_Pol));

So I put two memset in the loan.cpp and the new loans seem to be good
now. Attached is new exe. When you get the chance, please test it and
validate if the reports are not affected by the continuous running of
BMSim.